requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scikit-learn>=1.3.0
umap-learn>=0.5.3
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.65.0
scipy>=1.10.0
ripser>=0.6.4
persim>=0.3.1
requests>=2.28.0
h5py>=3.8.0
